chore: extract dedicated useArrowNavigation hook
